+++ builds/4chan-X-noupdate.user.js
@@ -1,6 +1,6 @@
 // ==UserScript==
 // @name         4chan X
-// @version      1.14.16.6
+// @version      1.14.16.7
 // @minGMVer     1.14
 // @minFFVer     26
 // @namespace    4chan-X
@@ -215,7 +215,7 @@ docSet = function() {
 };
 
 g = {
-  VERSION:   '1.14.16.6',
+  VERSION:   '1.14.16.7',
   NAMESPACE: '4chan X.',
   sites:     Object.create(null),
   boards:    Object.create(null)
@@ -454,7 +454,7 @@ Config = (function() {
       filesize: '',
       MD5: ''
     },
-    sauces: "# Known filename formats:\nhttps://www.pixiv.net/member_illust.php?mode=medium&illust_id=%$1;regexp:/^(\\d+)_p\\d+/\nhttps://www.deviantart.com/gallery/#/d%$1%$2;regexp:/^\\w+_by_\\w+[_-]d([\\da-z]{6})\\b|^d([\\da-z]{6})-[\\da-z]{8}-/\nhttps://imgur.com/%$1;regexp:/^(?![a-zA-Z][a-z]{6})(?![A-Z]{7})(?!\\d{7})([\\da-zA-Z]{7})(?: \\(\\d+\\))?\\.\\w+$/\nhttps://flickr.com/photo.gne?id=%$1;regexp:/^(\\d+)_[\\da-f]{10}(?:_\\w)*\\b/\nhttps://www.facebook.com/photo.php?fbid=%$1;regexp:/^\\d+_(\\d+)_\\d+_[no]\\b/\n\n# Reverse image search:\nhttps://www.google.com/searchbyimage?image_url=%IMG&safe=off\nhttps://yandex.com/images/search?rpt=imageview&url=%IMG\n#//tineye.com/search?url=%IMG\n#//www.bing.com/images/search?q=imgurl:%IMG&view=detailv2&iss=sbi#enterInsights\n\n# Specialized reverse image search:\n//iqdb.org/?url=%IMG\nhttps://trace.moe/?auto&url=%IMG;text:wait\n#//3d.iqdb.org/?url=%IMG\n#//saucenao.com/search.php?url=%IMG\n\n# \"View Same\" in archives:\nhttp://eye.swfchan.com/search/?q=%name;types:swf\n#https://desuarchive.org/_/search/image/%sMD5/\n#https://archive.4plebs.org/_/search/image/%sMD5/\n#https://boards.fireden.net/_/search/image/%sMD5/\n#https://foolz.fireden.net/_/search/image/%sMD5/\n\n# Other tools:\n#http://exif.regex.info/exif.cgi?imgurl=%URL\n#//imgops.com/%URL;types:gif,jpg,png\n#//www.gif-explode.com/%URL;types:gif",
+    sauces: "# Known filename formats:\nhttps://www.pixiv.net/member_illust.php?mode=medium&illust_id=%$1;regexp:/^(\\d+)_p\\d+/\njavascript:void(open(\"https://www.deviantart.com/\"+%$1.replace(/_/g,\"-\")+\"/art/\"+parseInt(%$2,36)));regexp:/^\\w+_by_(\\w+)[_-]d([\\da-z]{6})\\b/\nhttps://imgur.com/%$1;regexp:/^(?![a-zA-Z][a-z]{6})(?![A-Z]{7})(?!\\d{7})([\\da-zA-Z]{7})(?: \\(\\d+\\))?\\.\\w+$/\nhttps://flickr.com/photo.gne?id=%$1;regexp:/^(\\d+)_[\\da-f]{10}(?:_\\w)*\\b/\nhttps://www.facebook.com/photo.php?fbid=%$1;regexp:/^\\d+_(\\d+)_\\d+_[no]\\b/\n\n# Reverse image search:\nhttps://www.google.com/searchbyimage?image_url=%IMG&safe=off\nhttps://yandex.com/images/search?rpt=imageview&url=%IMG\n#//tineye.com/search?url=%IMG\n#//www.bing.com/images/search?q=imgurl:%IMG&view=detailv2&iss=sbi#enterInsights\n\n# Specialized reverse image search:\n//iqdb.org/?url=%IMG\nhttps://trace.moe/?auto&url=%IMG;text:wait\n#//3d.iqdb.org/?url=%IMG\n#//saucenao.com/search.php?url=%IMG\n\n# \"View Same\" in archives:\nhttp://eye.swfchan.com/search/?q=%name;types:swf\n#https://desuarchive.org/_/search/image/%sMD5/\n#https://archive.4plebs.org/_/search/image/%sMD5/\n#https://boards.fireden.net/_/search/image/%sMD5/\n#https://foolz.fireden.net/_/search/image/%sMD5/\n\n# Other tools:\n#http://exif.regex.info/exif.cgi?imgurl=%URL\n#//imgops.com/start?url=%URL;types:gif,jpg,png\n#//www.gif-explode.com/%URL;types:gif",
     FappeT: {
       werk: false
     },
@@ -13438,6 +13438,11 @@ Settings = (function() {
           set('archiveLists', data['archiveLists'].replace('https://mayhemydg.github.io/archives.json/archives.json', 'https://nstepien.github.io/archives.json/archives.json'));
         }
       }
+      if (compareString < '00001.00014.00016.00007') {
+        if (data['sauces'] != null) {
+          set('sauces', data['sauces'].replace(/https:\/\/www\.deviantart\.com\/gallery\/#\/d%\$1%\$2;regexp:\/\^\\w\+_by_\\w\+\[_-\]d\(\[\\da-z\]\{6\}\)\\b\|\^d\(\[\\da-z\]\{6\}\)-\[\\da-z\]\{8\}-\//g, 'javascript:void(open("https://www.deviantart.com/"+%$1.replace(/_/g,"-")+"/art/"+parseInt(%$2,36)));regexp:/^\\w+_by_(\\w+)[_-]d([\\da-z]{6})\\b/').replace(/\/\/imgops\.com\/%URL/g, '//imgops.com/start?url=%URL'));
+        }
+      }
       return changes;
     },
     loadSettings: function(data, cb) {
@@ -16803,7 +16808,11 @@ Embedding = (function() {
           case 200:
           case 304:
             text = service.text(req.response, uid);
-            Embedding.cache.set(data, text);
+            if (typeof text === 'string') {
+              Embedding.cache.set(data, text);
+            } else {
+              text = link.textContent;
+            }
             break;
           case 404:
             text = "Not Found";
@@ -16950,10 +16959,12 @@ Embedding = (function() {
         key: 'Gfycat',
         regExp: /^\w+:\/\/(?:www\.)?gfycat\.com\/(?:iframe\/)?(\w+)/,
         el: function(a) {
-          var div;
-          return div = $.el('iframe', {
-            src: "//gfycat.com/iframe/" + a.dataset.uid
+          var el;
+          el = $.el('iframe', {
+            src: "//gfycat.com/ifr/" + a.dataset.uid
           });
+          el.setAttribute("allowfullscreen", "true");
+          return el;
         }
       }, {
         key: 'Gist',
@@ -17005,11 +17016,11 @@ Embedding = (function() {
         }
       }, {
         key: 'LiveLeak',
-        regExp: /^\w+:\/\/(?:\w+\.)?liveleak\.com\/.*\?.*i=(\w+)/,
+        regExp: /^\w+:\/\/(?:\w+\.)?liveleak\.com\/.*\?.*[tif]=(\w+)/,
         el: function(a) {
           var el;
           el = $.el('iframe', {
-            src: "https://www.liveleak.com/ll_embed?i=" + a.dataset.uid
+            src: "https://www.liveleak.com/e/" + a.dataset.uid
           });
           el.setAttribute("allowfullscreen", "true");
           return el;
@@ -17115,14 +17126,14 @@ Embedding = (function() {
       }, {
         key: 'Openings.moe',
         regExp: /^\w+:\/\/openings.moe\/\?video=([^.&=]+)/,
-        style: 'max-width: 80vw; max-height: 80vh;',
+        style: 'width: 1280px; height: 720px; max-width: 80vw; max-height: 80vh;',
         el: function(a) {
-          return $.el('video', {
-            controls: true,
-            preload: 'auto',
-            src: "//openings.moe/video/" + a.dataset.uid + ".webm",
-            loop: true
+          var el;
+          el = $.el('iframe', {
+            src: "https://openings.moe/?video=" + a.dataset.uid
           });
+          el.setAttribute("allowfullscreen", "true");
+          return el;
         }
       }, {
         key: 'Pastebin',
@@ -18744,7 +18755,7 @@ FileInfo = (function() {
       n: function() {
         var fullname, shortname;
         fullname = this.file.name;
-        shortname = g.SITE.Build.shortFilename(this.file.name, this.isReply);
+        shortname = SW.yotsuba.Build.shortFilename(this.file.name, this.isReply);
         if (fullname === shortname) {
           return {innerHTML: E(fullname)};
         } else {